Convert simp_config.rb to class
This is a quick, dirty, and direct conversion from `simp_config.rb` to a
Ruby class.

Also: Fixes load path
